requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.0
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.0
python-dateutil==2.8.2
gunicorn==21.0.0psycopg2-binary==2.9.9